Struct ethcore_dapps::ServerBuilder

pub struct ServerBuilder {
    // some fields omitted
}

Webapps HTTP+RPC server build.

Methods

impl ServerBuilder

fn new(dapps_path: String, registrar: Arc<ContractClient>) -> Self

Construct new dapps server

fn with_sync_status(&mut self, status: Arc<SyncStatus>)

Change default sync status.

fn with_signer_port(&mut self, signer_port: Option<u16>)

Change default signer port.

fn start_unsecured_http(&self, addr: &SocketAddr, hosts: Option<Vec<String>>) -> Result<Server, ServerError>

Asynchronously start server with no authentication, returns result with Server handle on success or an error.

fn start_basic_auth_http(&self, addr: &SocketAddr, hosts: Option<Vec<String>>, username: &str, password: &str) -> Result<Server, ServerError>

Asynchronously start server with HTTP Basic Authentication, return result with Server handle on success or an error.