Struct ethcore_dapps::ServerBuilder

pub struct ServerBuilder<T: Fetch = FetchClient> { /* fields omitted */ }

Webapps HTTP+RPC server build.

Methods

impl ServerBuilder

fn new(dapps_path: String,
registrar: Arc<ContractClient>,
remote: Remote)
-> Self

Construct new dapps server

impl<T: Fetch> ServerBuilder<T>

fn fetch<X: Fetch>(self, fetch: X) -> ServerBuilder<X>

Set a fetch client to use.

fn sync_status(self, status: Arc<SyncStatus>) -> Self

Change default sync status.

fn web_proxy_tokens(self, tokens: Arc<WebProxyTokens>) -> Self

Change default web proxy tokens validator.

fn signer_address(self, signer_address: Option<(String, u16)>) -> Self

Change default signer port.

fn start_unsecured_http(self,
addr: &SocketAddr,
hosts: Option<Vec<String>>)
-> Result<Server, ServerError>

Asynchronously start server with no authentication, returns result with Server handle on success or an error.

fn start_basic_auth_http(self,
addr: &SocketAddr,
hosts: Option<Vec<String>>,
username: &str,
password: &str)
-> Result<Server, ServerError>

Asynchronously start server with HTTP Basic Authentication, return result with Server handle on success or an error.